refactor: using higher-level component imports
